+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.test.ts
@@ -1,3 +1,4 @@
+import { KeyringTypes } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 import type nock from 'nock';
 
@@ -762,6 +763,7 @@ describe('user-storage/user-storage-controller - syncInternalAccountsWithUserSto
     await controller.syncInternalAccountsWithUserStorage();
 
     expect(mockSyncInternalAccountsWithUserStorage).toHaveBeenCalled();
+    expect(controller.state.hasAccountSyncingSyncedAtLeastOnce).toBe(true);
   });
 });
 
@@ -966,6 +968,11 @@ describe('user-storage/user-storage-controller - account syncing edge cases', ()
       messenger: messengerMocks.messenger,
     });
 
+    await controller.setIsBackupAndSyncFeatureEnabled(
+      BACKUPANDSYNC_FEATURES.accountSyncing,
+      false,
+    );
+
     const mockSetStorage = jest.spyOn(controller, 'performSetStorage');
 
     // Create mock account
@@ -975,6 +982,9 @@ describe('user-storage/user-storage-controller - account syncing edge cases', ()
       metadata: {
         name: 'Test',
         nameLastUpdatedAt: Date.now(),
+        keyring: {
+          type: KeyringTypes.hd,
+        },
       },
     } as InternalAccount;
 
@@ -1008,6 +1018,9 @@ describe('user-storage/user-storage-controller - snap handling', () => {
 
     await expect(controller.getStorageKey()).rejects.toThrow(
       '#snapSignMessage - unable to call snap, wallet is locked',
+    );
+    await expect(controller.listEntropySources()).rejects.toThrow(
+      'listEntropySources - unable to list entropy sources, wallet is locked',
     );
   });